Privcand: Created db at first RPC instead of socket creation, fix discard/delete Privcand: On non-conflict, merge at update Change int to size_t in cxvec_append and related functions

+++ apps/backend/clixon_backend_plugin.h
@@ -60,12 +60,12 @@ typedef struct {
     cxobj     *td_src;      /* Source database xml tree */
     cxobj     *td_target;   /* Target database xml tree */
     cxobj    **td_dvec;     /* Delete xml vector */
-    int        td_dlen;     /* Delete xml vector length */
+    size_t     td_dlen;     /* Delete xml vector length */
     cxobj    **td_avec;     /* Add xml vector */
-    int        td_alen;     /* Add xml vector length */
+    size_t     td_alen;     /* Add xml vector length */
     cxobj    **td_scvec;    /* Source changed xml vector */
     cxobj    **td_tcvec;    /* Target changed xml vector */
-    int        td_clen;     /* Changed xml vector length */
+    size_t     td_clen;     /* Changed xml vector length */
 } transaction_data_t;
 
 /*! Pagination userdata
